+++ packages/moonstone/CHANGELOG.md
@@ -7,8 +7,8 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Added
 
 - `moonstone/VirtualList.VirtualList` and `moonstone/VirtualList.VirtualGridList` `role="list"`
-- `moonstone/VirtualList`, `moonstone/VirtualGridList` and `moonstone/Scroller` props 'horizontalScrollbarNextButtonAriaLabel', 'horizontalScrollbarPreviousButtonAriaLabel', 'verticalScrollbarNextButtonAriaLabel', and 'verticalScrollbarPreviousButtonAriaLabel' to configure the aria-label set on scroll buttons in the scrollbars
-- `moonstone/Panels` property `closeButtonAriaLabel` to configure the label set on application close button 
+- `moonstone/VirtualList`, `moonstone/VirtualGridList` and `moonstone/Scroller` props `scrollRightAriaLabel`, `scrollLeftAriaLabel`, `scrollDownAriaLabel`, and `scrollUpAriaLabel` to configure the aria-label set on scroll buttons in the scrollbars
+- `moonstone/Panels` property `closeButtonAriaLabel` to configure the label set on application close button
 
 ### Changed